requirements.txt
torch
pandas
numpy
tqdm
rdkit-pypi
requests
